+++ auto_route/example/lib/mobile/screens/profile/routes.dart
@@ -1,0 +1,15 @@
+import 'package:auto_route/annotations.dart';
+import 'package:auto_route/auto_route.dart';
+
+import 'my_books_page.dart';
+import 'profile_page.dart';
+
+const profileTab = AutoRoute(
+  path: 'profile',
+  name: 'ProfileTab',
+  page: EmptyRouterPage,
+  children: [
+    AutoRoute(path: '', page: ProfilePage),
+    AutoRoute(path: 'books', page: MyBooksPage),
+  ],
+);