+++ auto_route/lib/src/router/widgets/auto_router.dart
@@ -34,7 +34,8 @@ class AutoRouter extends StatefulWidget {
     var scope = StackRouterScope.of(context);
     assert(() {
       if (scope == null) {
-        throw FlutterError('AutoRouter operation requested with a context that does not include an AutoRouter.\n'
+        throw FlutterError(
+            'AutoRouter operation requested with a context that does not include an AutoRouter.\n'
             'The context used to retrieve the Router must be that of a widget that '
             'is a descendant of an AutoRouter widget.');
       }
@@ -90,7 +91,8 @@ class AutoRouterState extends State<AutoRouter> {
   }
 }
 
-typedef RoutesGenerator = List<PageRouteInfo> Function(BuildContext context, List<PageRouteInfo> routes);
+typedef RoutesGenerator = List<PageRouteInfo> Function(
+    BuildContext context, List<PageRouteInfo> routes);
 
 class _DeclarativeAutoRouter extends StatefulWidget {
   final RoutesGenerator onGenerateRoutes;
@@ -140,7 +142,8 @@ class _DeclarativeAutoRouterState extends State<_DeclarativeAutoRouter> {
       router: _controller,
       navigatorObservers: widget.navigatorObservers,
       didPop: (route) {
-        widget.onPopRoute?.call((route.settings as AutoRoutePage).routeData.route);
+        widget.onPopRoute
+            ?.call((route.settings as AutoRoutePage).routeData.route);
       },
     );
     return RoutingControllerScope(